Move `admin.helper.ts` to `noris.helper.ts`.
